perf(selectors): keep the wrapper collapse inside its budgeted closure
The extracted module grew the eager closure of three budgeted entries by one
module each -- interaction-targeting.ts 13 -> 14, selector-pipeline.ts 25 -> 26,
absence-observation-resolution.ts likewise -- and the eager-closure gate
ratchets that closure against the merge-base: an entry surface that drags more
of the repo onto the import path is a loading-shape regression whatever the
reason. The rule has exactly one consumer, so it now lives beside the
classification that asks it and is no longer an exported surface.

Its tests move to the owning module's test file and exercise
`classifyActionableTouchCandidates`, the boundary the command actually calls.
Each of the four refusals fails when its own guard is mutated: the hittability
condition, the 1 pt slack, the non-actionable-wrapper condition, and the
semantic-control condition.
